Refactor segment instructions
This commit refactors segment instructions to use/define a 64 bit
integer instead of a LLVM ptr type. This is required, as LLVM would
lower all pointers to 32 bit integers, and this would lose MTE tags when
storing them in WASM locals or to memory.

This commit is just a beginning of this refactoring, as only 32 bit
integer stores/loads are supported at this very moment. This change
allows us to compile our simple C demo. Future commits will continue
working on this.